03_spss/requirements.txt
tensorflow-gpu==1.12.0
librosa==0.7.1
tqdm==4.11.2
Unidecode==0.4.20
scipy
soundfile

